+++ backend/events/management/commands/consume.py
@@ -5,7 +5,6 @@ from django.core.management.base import BaseCommand
 
 from events import health
 from events.reactor import consume
-from events.reactor import resync
 
 logger = structlog.get_logger("events")
 
@@ -15,15 +14,11 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         health_service = health.HealthService()
-        logger.info("resync local rep")
-        # Init: resync all orchestrator assets
-        resync()
-
         logger.debug("starting consume loop")
-        # Consume rabbitmq messages indefinitely
+        # Consume grpc streams indefinitely
         while True:
             try:
                 consume(health_service)
             except Exception as e:
-                logger.exception("Error while consuming messages from the orchestrator RabbitMQ queue, will retry", e=e)
+                logger.exception("Error while consuming messages from the orchestrator gRPC streams, will retry", e=e)
                 time.sleep(5)