+++ backend/localrep/models/__init__.py
@@ -7,6 +7,7 @@ from .computetask import ComputeTaskInput
 from .computetask import ComputeTaskOutput
 from .datamanager import DataManager
 from .datasample import DataSample
+from .events import LastEvent
 from .model import Model
 from .organization import ChannelOrganization
 from .performance import Performance
@@ -24,4 +25,5 @@ __all__ = [
     "Model",
     "ChannelOrganization",
     "Performance",
+    "LastEvent",
 ]